Initial commit of CPU implementation.
This is the minimum necessary to run against the test roms (most of them fail currently).
There is a stubbed mmu that just operates on a flat memory space in intercepts the single serial register
used by the test roms for output. The program will only run against the single roms as there is no
rom banking implemented yet (needed for anything > 32 K). The first command line arg is used
to load the rom file, so they can just be dragged onto the executable to quickly run them.