Re-order machine cycle operation.
This reorders how cpu.cpp handles each machine cycle. It is reordered to make debugging
a bit easier to follow. This is achieved by making it so that the first thing in a cycle is the
memory access, and the very next thing is to decode the opcode. This provides a very
obvious location to check for interrupts, and also makes it simple to identify when the
PC can be modified to provide pseudo-debugging.

Unfortunately this also required some significant changes in how some operations
where encoded, notably the DATA pseudo register no longer exists, each memory
operation goes directly to and from a register.